fix(slack): finish tunnel install polish feedback
